% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbglm}
\alias{blbglm}
\title{Generalized linear regression with little bag of bootstraps}
\usage{
blbglm(formula, family, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{a formula of model.}

\item{family}{family for GLM}

\item{data}{a data frame.}

\item{m}{data splits, default = 10.}

\item{B}{bootstraps, default = 5000.}

\item{parallel}{logical value indicate use parallelization or not, default is not use.}
}
\value{
res A list including estimates and sigma.
}
\description{
Generalized linear regression with little bag of bootstraps
}
