% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm}
\alias{blblm}
\alias{blblm-package}
\title{Linear Regression with Little Bag of Bootstraps}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{a formula of model.}

\item{data}{a data frame.}

\item{m}{data splits, default = 10.}

\item{B}{bootstraps, default = 5000.}

\item{parallel}{logical value indicate use parallelization or not, default is not use.}
}
\value{
res A list including estimates and sigma.
}
\description{
Linear or Generalized Linear Regression with Little Bag of Bootstraps.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
\author{
\strong{Maintainer}: Zhanglin Shanggguan \email{zshangguan@ucdavis.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
